+++ packages/apps-config/src/endpoints/testingRelayPaseo.ts
@@ -324,7 +324,7 @@ export const testParasPaseo: Omit<EndpointOption, 'teleport'>[] = [
     info: 'quantum-fusion-paseo',
     paraId: 4775,
     providers: {
-      'Quantum Fusion': 'wss://para-test.qfnetwork.xyz'
+      // 'Quantum Fusion': 'wss://para-test.qfnetwork.xyz' // https://github.com/polkadot-js/apps/issues/11745
     },
     text: 'Quantum Fusion (Paseo)',
     ui: {
@@ -387,7 +387,7 @@ export const testParasPaseo: Omit<EndpointOption, 'teleport'>[] = [
     info: 'paseoXode',
     paraId: 4607,
     providers: {
-      XodeCommunity: 'wss://testrpcnodea01.xode.net/aRoyklGrhl9m2LlhX8NP/rpc'
+      XodeCommunity: 'wss://paseo-rpcnode.xode.net'
     },
     text: 'Xode',
     ui: {